# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Error": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TypeError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUser": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"created_at": {
						"anyOf": [
							{
								"type": "number",
								"format": "double"
							},
							{
								"type": "string"
							}
						]
					},
					"updated_at": {
						"anyOf": [
							{
								"type": "number",
								"format": "double"
							},
							{
								"type": "string"
							}
						]
					}
				},
				"required": [
					"id",
					"name",
					"email",
					"phone",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"error": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Error"
							},
							{
								"$ref": "#/components/schemas/TypeError"
							},
							{}
						]
					},
					"data": {
						"items": {
							"anyOf": [
								{
									"$ref": "#/components/schemas/IUser"
								},
								{}
							]
						},
						"type": "array"
					},
					"docs": {}
				},
				"type": "object",
				"additionalProperties": false
			},
			"IUserFilter": {
				"properties": {
					"qurey": {
						"type": "string"
					},
					"sortByEnumValue": {
						"type": "string"
					},
					"isAscending": {
						"type": "boolean"
					},
					"startDate": {
						"type": "string"
					},
					"endDate": {
						"type": "string"
					}
				},
				"required": [
					"sortByEnumValue"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IEnum": {
				"properties": {
					"text": {
						"type": "string"
					},
					"value": {
						"type": "string"
					}
				},
				"required": [
					"text",
					"value"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "user-crud",
		"version": "1.1.1",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/v1/user/create-update": {
			"post": {
				"operationId": "Create_update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUserResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUser"
							}
						}
					}
				}
			}
		},
		"/api/v1/user/filter": {
			"post": {
				"operationId": "FilterUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUserResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserFilter"
							}
						}
					}
				}
			}
		},
		"/api/v1/user/delete": {
			"post": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUserResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"id": {
										"type": "string"
									}
								},
								"required": [
									"id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/v1/user/list": {
			"get": {
				"operationId": "GetAllUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUserResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/v1/enums": {
			"get": {
				"operationId": "GetEnums",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"UserSorting": {
											"items": {
												"$ref": "#/components/schemas/IEnum"
											},
											"type": "array"
										}
									},
									"required": [
										"UserSorting"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Enums"
				],
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}